Move dialogs logic to renderer process
